+++ src/main/java/programmers/org/voucher/repository/util/sqlBuilder/UpdateBuilder.java
@@ -1,7 +1,7 @@
 package programmers.org.voucher.repository.util.sqlBuilder;
 
-import programmers.org.voucher.repository.util.statement.Set;
-import programmers.org.voucher.repository.util.statement.Update;
+import programmers.org.voucher.repository.util.statement.update.Set;
+import programmers.org.voucher.repository.util.statement.update.Update;
 import programmers.org.voucher.repository.util.statement.Where;
 
 import static programmers.org.voucher.repository.util.constant.Symbol.BLANK;
